requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cassandra-driver==3.28.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.104.0
geomet==0.2.1.post1
greenlet==3.0.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
pyasn1==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.7
